requirements.txt
# Core LangChain and dependencies
langchain==0.3.20
langchain-core==0.3.45
langchain-community==0.3.19
langchain-text-splitters==0.3.6
langchain-google-genai==2.1.0

# Embedding & vector storage
faiss-cpu==1.10.0
sentence-transformers==2.6.1
huggingface_hub==0.23.3

# Google Generative AI
google-ai-generativelanguage==0.6.17

# Utilities
PyPDF2==3.0.1
pandas==2.2.3
tqdm==4.66.4
python-dotenv==1.0.1

# UI
streamlit==1.43.2

langchain-openai
